% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_model}
\alias{ridge_model}
\title{Ridge Regression}
\usage{
ridge_model(X, y)
}
\arguments{
\item{X}{Matrix of predictors. Each row represents an observation, and each column represents a predictor.
can be a mix of continuous, discrete, and binary predictors.}

\item{y}{Response variable. For binary outcomes, it should be a binary vector (0 or 1). For continuous outcomes, it should be a numeric vector.}
}
\value{
A list containing the fitted ridge regression model and predicted values.
}
\description{
Ridge Regression
}
\examples{
# Example usage:
X <- matrix(rnorm(100 * 2), ncol = 2)
y <- rnorm(100)
result <- ridge_model(X, y)

}
